Removed the test paragraph moved the home page url
I removed the test paragraph inside the ul list, I just wanted to test if it is possible to include such an element inside a list.
Moved the home page(index) link to the bottom of the page as it looks better and users are more likely to want to return to the
home page after they've went through the recipe.